chore: make Karma peerDependency
It makes no sense to install two instances of Karma, if the developer wants to use both Karma through grunt-karma and Karma directly.

Also, since Karma 0.9, developers need to install plugins and it is very hard to load plugins if Karma was installed as a dependency of grunt-karma.

The recommended way is to put everything into the `package.json`:
```json
{
  "devDependencies": {
    "karma": "~0.10",
    "grunt": "~0.4",
    "grunt-karma": "*",
    "karma-mocha": "*"
  }
}
```

Closes #13