Listen for SIGINT and stop the process.
Karma listens for SIGINT and disconnects all servers if interrupted.
As a side-effect, interrupting the current process causes Karma to stop
but nothing else. Any pending Grunt tasks will still run.

Since Grunt does not listen to SIGINT itself, the expected behaviour
is that the process exits with code 130. We simulate this by listening
for SIGINT and calling process.exit() when Karma finished.