chore: make travis and local 'npm test' do the same
It was confusing to be that for CI we run 'grunt [default]',
but during development we'd use 'npm test' (set to 'grunt test',
which didn't include 'eslint').

To make sure we get the right feedback locally, make sure eslint
is included in both.

Also, while at it, let Travis invoke 'grunt' via 'npm test', so
that it automatically has local grunt (devDependencies) in its
PATH and doesn't need it installed a second time globally.